PHP to Javascript change and visual update.
So this project was originally intended to showcase some PHP knowledge,
but it turns out that a lot of the calls were much more appropriate for
jQuery/JSON, so I replaced the PHP api calls. Pretty happy with how
it's looking so far, but there's still a lot of work to be done. Next
plans are to get related videos showing up correctly.